moving objects around into their own files
