+++ boardgame_v1/boardgame/coord.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Coord:
+    rank: int
+    file: int
