+++ boardgame_v1/boardgame/direction.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class Direction(Enum):
+    UP = 1
+    RIGHT = 2
