+++ boardgame_v1/boardgame/game.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+from boardgame_v1.boardgame.board import Board
+
+
+@dataclass
+class Game:
+    board: Board
