+++ boardgame_v1/boardgame/player.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Player:
+    name: str
