+++ boardgame_v1/boardgame/vector.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+from boardgame_v1.boardgame.direction import Direction
+
+
+@dataclass
+class Vector:
+    direction: Direction
+    scalar: int
